LayoutTests/media/controls-right-click-on-timebar-expected.txt

Test that right clicking on the timebar does not cause a seek.

RUN(video.autoplay = true)
RUN(video.src = 'content/test.mp4')
EVENT(playing)
END OF TEST
